it works (just) but is decently slower than the non-monadic version

+++ src/Probability.hs
@@ -3,8 +3,8 @@ module Probability(module Probability.Internal, dice, normalise) where
     import Data.Bifunctor
     import Data.Ratio
 
-    dice :: Integer -> Probability Integer
-    dice n = P  . zip [1 .. n] . repeat $ 1%n
+    dice :: Int -> Probability Int
+    dice n = P  . zip [1 .. n] . repeat $ 1%fromIntegral n
 
     normalise :: Probability a -> Probability a
     normalise (P x) = P (map (second (/s)) x)
